+++ gpkit/__init__.py
@@ -90,8 +90,7 @@ from .nomials import Monomial, Posynomial, Signomial
 from .variables import Variable, VectorVariable, ArrayVariable
 from .varkey import VarKey
 from .posyarray import PosyArray
-from .newModel import Model
-from .signomial_program import SignomialProgram
+from .model import Model
 from .user_scripts import composite_objective, link
 from .user_scripts import make_feasibility_gp, closest_feasible_point
