+++ gpkit/solution_array.py
