+++ gpkit/tests/run_tests.py
@@ -20,11 +20,11 @@ TESTS += t_constraints.TESTS
 from gpkit.tests import t_posy_array
 TESTS += t_posy_array.TESTS
 
-from gpkit.tests import t_geometric_program
-TESTS += t_geometric_program.TESTS
+#from gpkit.tests import t_geometric_program
+#TESTS += t_geometric_program.TESTS
 
-from gpkit.tests import t_gp_solution_array
-TESTS += t_gp_solution_array.TESTS
+#from gpkit.tests import t_gp_solution_array
+#TESTS += t_gp_solution_array.TESTS
 
 
 import gpkit
